add test to coordinate view
