+++ src/landscape/coordinate_point/mod.rs
@@ -70,3 +70,66 @@ impl std::ops::Mul<f64> for CoordinatePoint {
         Self(self.0 * rhs, self.1 * rhs)
     }
 }
+
+#[derive(Default, Debug, Deserialize, Serialize)]
+pub struct Coordinates(pub Vec<CoordinatePoint>);
+
+impl Coordinates {
+    pub fn view_for_landscape_position(
+        &self,
+        landscape_position: &CoordinatePoint,
+    ) -> CoordinateView {
+        CoordinateView::for_landscape_position(self, landscape_position)
+    }
+}
+
+/// The CoordinateView is constructed to access a list of coordinates
+/// relative to a central point (landscape_position). All coordinates
+/// will be transformed to be relative to this point and all y will be
+/// inverted to align with Bevy's inane coordinate system and make
+/// spawning things easy.
+#[derive(Default, Debug, Deserialize, Serialize)]
+pub struct CoordinateView {
+    pub list: Vec<Vec2>,
+    pub center: Vec2,
+    pub max_x: f32,
+    pub max_y: f32,
+    pub min_x: f32,
+    pub min_y: f32,
+}
+
+impl CoordinateView {
+    pub fn for_landscape_position(
+        coordinates: &Coordinates,
+        landscape_position: &CoordinatePoint,
+    ) -> Self {
+        let list: Vec<Vec2> = coordinates
+            .0
+            .iter()
+            .map(|coordinate| {
+                let coordinates: Vec2 = (*coordinate - *landscape_position).into();
+                coordinates * Vec2::new(1.0, -1.0)
+            })
+            .collect();
+
+        let x: Vec<f32> = list.iter().map(|e| e.x).collect();
+        let y: Vec<f32> = list.iter().map(|e| e.y).collect();
+
+        let max_x = *x.iter().max_by(|a, b| a.total_cmp(b)).unwrap();
+        let max_y = *y.iter().max_by(|a, b| a.total_cmp(b)).unwrap();
+        let min_x = *x.iter().min_by(|a, b| a.total_cmp(b)).unwrap();
+        let min_y = *y.iter().min_by(|a, b| a.total_cmp(b)).unwrap();
+
+        let center = Vec2::new(max_x + min_x, max_y + min_y) / 2.0;
+        let list: Vec<Vec2> = list.iter().map(|item| *item - center).collect();
+
+        Self {
+            min_x,
+            max_x,
+            min_y,
+            max_y,
+            list,
+            center,
+        }
+    }
+}
