+++ src/landscape/open_street_map/osm_data/mod.rs
@@ -4,7 +4,10 @@ mod tests;
 mod helpers;
 
 use super::{Path, PathId};
-use crate::{landscape::CoordinatePoint, train::Direction};
+use crate::{
+    landscape::{coordinate_point::Coordinates, CoordinatePoint},
+    train::Direction,
+};
 use bevy::prelude::*;
 use helpers::*;
 use proj::Proj;
@@ -36,65 +39,6 @@ pub struct BuildingData {
 pub struct AreaData {
     pub area_type: AreaType,
     pub coordinates: Coordinates,
-}
-
-#[derive(Default, Debug, Deserialize, Serialize)]
-pub struct Coordinates(Vec<CoordinatePoint>);
-
-#[derive(Default, Debug, Deserialize, Serialize)]
-pub struct CoordinateView {
-    pub list: Vec<Vec2>,
-    pub center: Vec2,
-    pub max_x: f32,
-    pub max_y: f32,
-    pub min_x: f32,
-    pub min_y: f32,
-}
-
-impl CoordinateView {
-    pub fn for_landscape_position(
-        coordinates: &Coordinates,
-        landscape_position: &CoordinatePoint,
-    ) -> Self {
-        let list: Vec<Vec2> = coordinates
-            .0
-            .iter()
-            .map(|coordinate| {
-                let coordinates: Vec2 = (*coordinate - *landscape_position).into();
-                coordinates * Vec2::new(1.0, -1.0)
-            })
-            .collect();
-
-        // normalize so that first element is zero
-        let x: Vec<f32> = list.iter().map(|e| e.x).collect();
-        let y: Vec<f32> = list.iter().map(|e| e.y).collect();
-
-        let max_x = *x.iter().max_by(|a, b| a.total_cmp(b)).unwrap();
-        let max_y = *y.iter().max_by(|a, b| a.total_cmp(b)).unwrap();
-        let min_x = *x.iter().min_by(|a, b| a.total_cmp(b)).unwrap();
-        let min_y = *y.iter().min_by(|a, b| a.total_cmp(b)).unwrap();
-
-        let center = Vec2::new(max_x + min_x, max_y + min_y) / 2.0;
-        let list: Vec<Vec2> = list.iter().map(|item| *item - center).collect();
-
-        Self {
-            min_x,
-            max_x,
-            min_y,
-            max_y,
-            list,
-            center,
-        }
-    }
-}
-
-impl Coordinates {
-    pub fn view_for_landscape_position(
-        &self,
-        landscape_position: &CoordinatePoint,
-    ) -> CoordinateView {
-        CoordinateView::for_landscape_position(self, landscape_position)
-    }
 }
 
 #[derive(Debug, PartialEq, Deserialize, Serialize)]
